Move issue tests to separated package
